+++ xmhf/src/xmhf-core/include/arch/x86_64/xmhf-baseplatform-arch-x86_64.h
@@ -326,7 +326,7 @@ void xmhf_baseplatform_arch_x86_64_wakeupAPs(void);
 void xmhf_baseplatform_arch_x86_64_reboot(void);
 
 //get the physical address of the root system description pointer (rsdp)
-u32 xmhf_baseplatform_arch_x86_64_acpi_getRSDP(ACPI_RSDP *rsdp);
+uintptr_t xmhf_baseplatform_arch_x86_64_acpi_getRSDP(ACPI_RSDP *rsdp);
 
 //PCI subsystem initialization
 void xmhf_baseplatform_arch_x86_64_pci_initialize(void);
